♻️refactor: 공통으로 사용하는 타입 분리
