+++ src/components/reservationList/reservations.types.ts
@@ -1,0 +1,1 @@
+export type StatusType = 'pending' | 'confirmed' | 'declined' | 'canceled' | 'completed';
